Sentiment_Analysis/requirements.txt
numpy==1.26.4
pandas==2.1.4
tensorflow-cpu==2.17.0
matplotlib==3.7.1
scikit-learn==1.0.2